tweaks

+++ prehension/visualization/session_data_visualization.py
@@ -27,6 +27,8 @@ import shutil
 import glob
 import pandas as pd
 import matplotlib.colors as mcolors
+import matplotlib
+matplotlib.use('Agg')  # Use non-interactive backend
 import matplotlib.pyplot as plt
 import numpy as np
 
@@ -458,7 +460,7 @@ class SessionWrapper:
             # logic:
             src_exists = os.path.exists(src_session_dir)
             if not src_exists:  # This should always exist
-                raise Exception(f"Source session directory {src_session_dir} does not exist")
+                raise FileNotFoundError(f"Source session directory {src_session_dir} does not exist")
 
             # Check if dst_session_dir exists
             already_transferred = False
@@ -467,17 +469,26 @@ class SessionWrapper:
                 if src_contents.issubset(dst_contents):
                     already_transferred = True
 
+            if already_transferred:
+                rs('Folder already transfered')
+                return
+
             # To keep track of transfer status
             transferred = False
 
             if not already_transferred or overwrite:
                 # rs('transferring')
                 # copy src to dst_parent_dir but remove dst first
+
+
                 if os.path.exists(dst_session_dir):
-                    try:
-                        shutil.rmtree(dst_session_dir, ignore_errors=False)
-                    except:
-                        ws(f"failed to remove {dst_session_dir}")
+                    ws('Warning, dst session directory already exists, overwriting...')
+
+                # if os.path.exists(dst_session_dir):
+                #     try:
+                #         shutil.rmtree(dst_session_dir, ignore_errors=False)
+                #     except:
+                #         ws(f"failed to remove {dst_session_dir}")
 
                 try:
                     shutil.move(src_session_dir, dst_parent_dir)
@@ -491,13 +502,13 @@ class SessionWrapper:
                 src_contents, dst_contents = _folder_contents_set(src_session_dir, dst_session_dir)
                 assert src_contents.issubset(dst_contents)
 
-            # now remove src locally because we should have already transferred
-            if os.path.exists(src_session_dir):
-                assert os.path.exists(dst_session_dir), "dir not found: {dst_session_dir}"
-                try:
-                    shutil.rmtree(src_session_dir, ignore_errors=False)
-                except:
-                    ws(f"failed to remove {src_session_dir}")
+                # now remove src locally because we should have already transferred
+                if os.path.exists(src_session_dir):
+                    assert os.path.exists(dst_session_dir), "dir not found: {dst_session_dir}"
+                    try:
+                        shutil.rmtree(src_session_dir, ignore_errors=False)
+                    except:
+                        ws(f"failed to remove {src_session_dir}")
 
         # Transfer from experiment to training server
         _move_helper(self.raw_ss, raw_training_server)
@@ -555,7 +566,10 @@ class SessionWrapper:
             # Extract the set number if available, otherwise return -1
             set_number = int(match.group(2)) if match.group(2) else -1
 
-        return current_date, set_number
+            return current_date, set_number
+
+        else:
+            raise ValueError(f'Could not parse date from folder name: {folder}')
 
     def plot_force_trace(self):
         tp_path = os.path.join(self.proc_ss, "timepoints.csv")
@@ -584,6 +598,7 @@ class SessionWrapper:
 
         # Return and THEN bind results to trial object
         return (times, forces_summed)
+
 
     def plot_force_traces(
             self,
